requirements.txt
backcall==0.1.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.10
django-easy-select2==1.5.6
django-extensions==2.2.9
django-filter==1.1.0
django-import-export==2.0.2
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-shell-plus==1.1.7
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
et-xmlfile==1.0.1
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.16.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sqlparse==0.3.1
tablib==1.1.0
traitlets==4.3.3
wcwidth==0.1.9
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
